# docs/swagger.yaml
definitions:
  commons.DataTableColumn:
    properties:
      name:
        type: string
      orderable:
        type: boolean
      searchValue:
        type: string
      searchable:
        type: boolean
    type: object
  commons.DataTableFilter:
    properties:
      field:
        type: string
      operator:
        type: string
      value:
        type: string
    type: object
  commons.DataTableOrder:
    properties:
      direction:
        type: string
      field:
        type: string
    type: object
  commons.DataTableRequest:
    properties:
      columns:
        items:
          $ref: '#/definitions/commons.DataTableColumn'
        type: array
      filters:
        items:
          $ref: '#/definitions/commons.DataTableFilter'
        type: array
      orders:
        items:
          $ref: '#/definitions/commons.DataTableOrder'
        type: array
      page:
        type: integer
      search:
        type: string
      size:
        type: integer
    type: object
...
  dto.NewProductDto:
    properties:
      name:
        type: string
      product_category_id:
        type: string
    required:
    - name
    - product_category_id
    type: object
...
  models.Product:
    properties:
      approved_at:
        type: string
      approved_by:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      entity_id:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_default:
        type: boolean
      is_locked:
        type: boolean
      name:
        type: string
      organization_id:
        type: string
      owner_id:
        type: string
      product_category_id:
        type: string
      submitted_at:
        type: string
      submitted_by:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  models.ProductDetail:
    properties:
      approved_at:
        type: string
      approved_by:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      entity_id:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_default:
        type: boolean
      is_locked:
        type: boolean
      item_name:
        type: string
      organization_id:
        type: string
      owner_id:
        type: string
      product_id:
        type: string
      service_type_id:
        type: string
      submitted_at:
        type: string
      submitted_by:
        type: string
      uom_id:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.ServiceType:
    properties:
      approved_at:
        type: string
      approved_by:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      entity_id:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_default:
        type: boolean
      is_locked:
        type: boolean
      name:
        type: string
      organization_id:
        type: string
      owner_id:
        type: string
      submitted_at:
        type: string
      submitted_by:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /product/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Product
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.NewProductDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "204":
          description: No Content
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Product
  /product/deleteById/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "204":
          description: No Content
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Product
  /product/detail/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Product Detail
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/models.ProductDetail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "204":
          description: No Content
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Product
  /product/getAll:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "204":
          description: No Content
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Product
  /product/getById/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "204":
          description: No Content
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Product
  /product/getPagination:
    post:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/commons.DataTableRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "204":
          description: No Content
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Product
  /product/update:
    patch:
      consumes:
      - application/json
      parameters:
      - description: product
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "204":
          description: No Content
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Product
...
  /service_type/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: ServiceType
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/models.ServiceType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "204":
          description: No Content
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Service Type
  /service_type/deleteById/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "204":
          description: No Content
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Service Type
  /service_type/getAll:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "204":
          description: No Content
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Service Type
  /service_type/getById/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "204":
          description: No Content
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Service Type
  /service_type/getPagination:
    post:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/commons.DataTableRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "204":
          description: No Content
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Service Type
  /service_type/update:
    patch:
      consumes:
      - application/json
      parameters:
      - description: serviceType
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/models.ServiceType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "204":
          description: No Content
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      tags:
      - Service Type
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"